swapped two functions between sbml_dfs core and sbml_dfs_utils and made the core function a method
