+++ src/tests/test_sbml_dfs_core.py
@@ -18,7 +18,6 @@ from napistu.constants import (
     BQB_DEFINING_ATTRS,
     BQB_DEFINING_ATTRS_LOOSE,
     BQB,
-    IDENTIFIERS,
 )
 from napistu.sbml_dfs_core import SBML_dfs
 
@@ -554,85 +553,48 @@ def test_remove_entity_data_nonexistent(sbml_dfs_w_data, caplog):
     sbml_dfs_w_data.validate()
 
 
-def test_filter_to_characteristic_species_ids():
-
-    species_ids_dict = {
-        SBML_DFS.S_ID: ["large_complex"] * 6
-        + ["small_complex"] * 2
-        + ["proteinA", "proteinB"]
-        + ["proteinC"] * 3
-        + [
-            "promiscuous_complexA",
-            "promiscuous_complexB",
-            "promiscuous_complexC",
-            "promiscuous_complexD",
-            "promiscuous_complexE",
-        ],
-        IDENTIFIERS.ONTOLOGY: ["complexportal"]
-        + ["HGNC"] * 7
-        + ["GO"] * 2
-        + ["ENSG", "ENSP", "pubmed"]
-        + ["HGNC"] * 5,
-        IDENTIFIERS.IDENTIFIER: [
-            "CPX-BIG",
-            "mem1",
-            "mem2",
-            "mem3",
-            "mem4",
-            "mem5",
-            "part1",
-            "part2",
-            "GO:1",
-            "GO:2",
-            "dna_seq",
-            "protein_seq",
-            "my_cool_pub",
-        ]
-        + ["promiscuous_complex"] * 5,
-        IDENTIFIERS.BQB: [BQB.IS]
-        + [BQB.HAS_PART] * 7
-        + [BQB.IS] * 2
-        + [
-            # these are retained if BQB_DEFINING_ATTRS_LOOSE is used
-            BQB.ENCODES,
-            BQB.IS_ENCODED_BY,
-            # this should always be removed
-            BQB.IS_DESCRIBED_BY,
-        ]
-        + [BQB.HAS_PART] * 5,
-    }
-
-    species_ids = pd.DataFrame(species_ids_dict)
-
-    characteristic_ids_narrow = sbml_dfs_core.filter_to_characteristic_species_ids(
-        species_ids,
-        defining_biological_qualifiers=BQB_DEFINING_ATTRS,
-        max_complex_size=4,
-        max_promiscuity=4,
+def test_get_characteristic_species_ids():
+    """
+    Test get_characteristic_species_ids function with both dogmatic and non-dogmatic cases.
+    """
+    # Create mock species identifiers data
+    mock_species_ids = pd.DataFrame(
+        {
+            "s_id": ["s1", "s2", "s3", "s4", "s5"],
+            "identifier": ["P12345", "CHEBI:15377", "GO:12345", "P67890", "P67890"],
+            "ontology": ["uniprot", "chebi", "go", "uniprot", "chebi"],
+            "bqb": [
+                "BQB_IS",
+                "BQB_IS",
+                "BQB_HAS_PART",
+                "BQB_HAS_VERSION",
+                "BQB_ENCODES",
+            ],
+        }
     )
 
-    EXPECTED_IDS = ["CPX-BIG", "GO:1", "GO:2", "part1", "part2"]
-    assert characteristic_ids_narrow[IDENTIFIERS.IDENTIFIER].tolist() == EXPECTED_IDS
+    # Create mock SBML_dfs object
+    class MockSBML_dfs:
+        def get_identifiers(self, entity_type):
+            return mock_species_ids
 
-    characteristic_ids_loose = sbml_dfs_core.filter_to_characteristic_species_ids(
-        species_ids,
-        # include encodes and is_encoded_by as equivalent to is
-        defining_biological_qualifiers=BQB_DEFINING_ATTRS_LOOSE,
-        max_complex_size=4,
-        # expand promiscuity to default value
-        max_promiscuity=20,
+    mock_sbml = MockSBML_dfs()
+
+    # Test dogmatic case (default)
+    expected_bqbs = BQB_DEFINING_ATTRS + [BQB.HAS_PART]  # noqa: F841
+    dogmatic_result = mock_sbml.get_characteristic_species_ids()
+    expected_dogmatic = mock_species_ids.query("bqb in @expected_bqbs")
+
+    pd.testing.assert_frame_equal(dogmatic_result, expected_dogmatic, check_like=True)
+
+    # Test non-dogmatic case
+    expected_bqbs = BQB_DEFINING_ATTRS_LOOSE + [BQB.HAS_PART]  # noqa: F841
+    non_dogmatic_result = mock_sbml.get_characteristic_species_ids(dogmatic=False)
+    expected_non_dogmatic = mock_species_ids.query("bqb in @expected_bqbs")
+
+    pd.testing.assert_frame_equal(
+        non_dogmatic_result, expected_non_dogmatic, check_like=True
     )
-
-    EXPECTED_IDS = [
-        "CPX-BIG",
-        "GO:1",
-        "GO:2",
-        "dna_seq",
-        "protein_seq",
-        "part1",
-        "part2",
-    ] + ["promiscuous_complex"] * 5
-    assert characteristic_ids_loose[IDENTIFIERS.IDENTIFIER].tolist() == EXPECTED_IDS
 
 
 def test_sbml_basic_functionality(test_data):
